fix: switch to BroadcastChannel
To support service workers, use BroadcastChannels where available
instead of postMessage.

Also supports BroadcastChannels in node worker threads. Attempts to
auto detect work threads vs cluster workers where the BroadcastChannel
API is not available.

Fixes #67
