+++ test/fixtures/worker-post-message.browser.ts
@@ -1,0 +1,3 @@
+export function postMessage (val: any) {
+  globalThis.postMessage(val)
+}
