+++ test/fixtures/worker-post-message.ts
@@ -1,0 +1,5 @@
+import worker from 'node:worker_threads'
+
+export function postMessage (val: any) {
+  worker.parentPort?.postMessage(val)
+}
