+++ test/node.ts
@@ -1,2 +1,3 @@
 import './index.test.js'
 import './cluster.test.js'
+import './worker-threads.test.js'
